Split the `composite-client` into several JS files
