+++ client-js/main/client/composite-client.js
@@ -20,15 +20,7 @@
 
 "use strict";
 
-import {Message} from 'google-protobuf';
-import ObjectToProto from "./object-to-proto";
-import {CommandHandlingError, CommandValidationError, SpineError} from "./errors";
-import {Status} from '../proto/spine/core/response_pb';
 import {Client} from "./client";
-import {TypedMessage} from "./typed-message";
-import {QueryRequest} from "./query-request";
-import {EventSubscriptionRequest, SubscriptionRequest} from "./subscribing-request";
-import {CommandRequest} from "./command-request";
 
 /**
  * A {@link Client} that delegates requests to case-specific client implementations.
@@ -132,222 +124,5 @@ export class CompositeClient extends Client {
    */
   post(command, onAck) {
     this._commanding.post(command, onAck);
-  }
-}
-
-/**
- * A client which performs entity state queries.
- *
- * @abstract
- */
-export class QueryingClient {
-
-  /**
-   * @param {!ActorRequestFactory} actorRequestFactory
-   *        a request factory to build requests to Spine server
-   */
-  constructor(actorRequestFactory) {
-    this._requestFactory = actorRequestFactory;
-  }
-
-  /**
-   * Creates a new query request.
-   *
-   * @param {!Class<? extends Message>} entityType the target entity type
-   * @param {!Client} client the client which initiated the request
-   * @return {QueryRequest} a new query request
-   */
-  select(entityType, client) {
-    return new QueryRequest(entityType, client, this._requestFactory);
-  }
-
-  /**
-   * Executes the given `Query` instance specifying the data to be retrieved from
-   * Spine server fulfilling a returned promise with an array of received objects.
-   *
-   * @param {!spine.client.Query} query a query instance to be executed
-   * @return {Promise<<T extends Message>[]>} a promise to be fulfilled with a list of Protobuf
-   *        messages of a given type or with an empty list if no entities matching given query
-   *        were found; rejected with a `SpineError` if error occurs
-   *
-   * @template <T> a Protobuf type of entities being the target of a query
-   */
-  read(query) {
-    throw new Error('Not implemented in abstract base.');
-  }
-
-  /**
-   * Creates a new query factory instance which can be further used for the `Query` creation.
-   *
-   * @return {QueryFactory}
-   */
-  newQuery() {
-    return this._requestFactory.query();
-  }
-}
-
-/**
- * A client which manages entity state and event subscriptions.
- *
- * @abstract
- */
-export class SubscribingClient {
-
-  /**
-   * @param {!ActorRequestFactory} actorRequestFactory
-   *        a request factory to build requests to Spine server
-   */
-  constructor(actorRequestFactory) {
-    this._requestFactory = actorRequestFactory;
-  }
-
-  /**
-   * Creates a new subscription request.
-   *
-   * @param {!Class<? extends Message>} type the target entity type
-   * @param {!Client} client the client that initiated the request
-   * @return {SubscriptionRequest} a new subscription request
-   */
-  subscribeTo(type, client) {
-    return new SubscriptionRequest(type, client, this._requestFactory);
-  }
-
-  /**
-   * Subscribes to a given topic which targets an entity type.
-   *
-   * @param {!spine.client.Topic} topic a topic to subscribe to
-   * @return {Promise<EntitySubscriptionObject<T extends Message>>} a subscription object
-   *
-   * @template <T> a Protobuf type of entities being the target of a subscription
-   */
-  subscribe(topic) {
-    throw new Error('Not implemented in abstract base.');
-  }
-
-  /**
-   * Creates a new event subscription request.
-   *
-   * @param {!Class<? extends Message>} type the target event type
-   * @param {!Client} client the client that initiated the request
-   * @return {EventSubscriptionRequest} a new event subscription request
-   */
-  subscribeToEvent(type, client) {
-    return new EventSubscriptionRequest(type, client, this._requestFactory);
-  }
-
-  /**
-   * Subscribes to the given topic which targets an event type.
-   *
-   * @param {!spine.client.Topic} topic a topic to subscribe to
-   * @return {Promise<EventSubscriptionObject>} a subscription object
-   */
-  subscribeToEvents(topic) {
-    throw new Error('Not implemented in abstract base.');
-  }
-
-  /**
-   * Returns a new topic factory instance which can be further used for the `Topic` creation.
-   *
-   * @return {TopicFactory}
-   */
-  newTopic() {
-    return this._requestFactory.topic();
-  }
-}
-
-const SUBSCRIPTIONS_NOT_SUPPORTED = 'Subscriptions are not supported.';
-
-/**
- * A {@link SubscribingClient} which does not create subscriptions.
- */
-export class NoOpSubscribingClient extends SubscribingClient {
-
-  constructor(actorRequestFactory) {
-    super(actorRequestFactory)
-  }
-
-  /**
-   * Always throws an error.
-   *
-   * @override
-   */
-  subscribe(topic) {
-    throw new Error(SUBSCRIPTIONS_NOT_SUPPORTED);
-  }
-
-  /**
-   * Always throws an error.
-   *
-   * @override
-   */
-  subscribeToEvents(topic) {
-    throw new Error(SUBSCRIPTIONS_NOT_SUPPORTED);
-  }
-}
-
-const _statusType = Status.typeUrl();
-
-/**
- * A client which posts commands.
- *
- * This class has a default implementation. Override it to change the behaviour.
- */
-export class CommandingClient {
-
-  constructor(endpoint, requestFactory) {
-    this._requestFactory = requestFactory;
-    this._endpoint = endpoint;
-  }
-
-  /**
-   * Creates a new command request.
-   *
-   * @param {!Message} commandMessage the command to send to the server
-   * @param {!Client} client the client which initiated the request
-   * @return {CommandRequest} a new command request
-   */
-  command(commandMessage, client) {
-    return new CommandRequest(commandMessage, client, this._requestFactory);
-  }
-
-  /**
-   * Posts a given command to the Spine server.
-   *
-   * @param {!spine.core.Command} command a Command sent to Spine server
-   * @param {!AckCallback} onAck a command acknowledgement callback
-   */
-  post(command, onAck) {
-    const cmd = TypedMessage.of(command);
-    this._endpoint.command(cmd)
-        .then(ack => this._onAck(ack, onAck))
-        .catch(error => {
-          onAck.onError(new CommandHandlingError(error.message, error));
-        });
-  }
-
-  _onAck(ack, onAck) {
-    const responseStatus = ack.status;
-    const responseStatusProto = ObjectToProto.convert(responseStatus, _statusType);
-    const responseStatusCase = responseStatusProto.getStatusCase();
-
-    switch (responseStatusCase) {
-      case Status.StatusCase.OK:
-        onAck.onOk();
-        break;
-      case Status.StatusCase.ERROR:
-        const error = responseStatusProto.getError();
-        const message = error.getMessage();
-        onAck.onError(error.hasValidationError()
-            ? new CommandValidationError(message, error)
-            : new CommandHandlingError(message, error));
-        break;
-      case Status.StatusCase.REJECTION:
-        onAck.onRejection(responseStatusProto.getRejection());
-        break;
-      default:
-        onAck.onError(
-            new SpineError(`Unknown response status case ${responseStatusCase}`)
-        );
-    }
   }
 }
